GLSL shaders for TileMap (#14)
This implements proper GLSL shaders and removes all reliance on the CPU for texture building. 0.1 was meant to just get something out there quickly for people to use before I did the research and had proper help with writing this part of the project.

Much of this work draws on the current way that Bevy's sprites are laid out with a custom mesh.